User commit

+++ src/main/java/com/example/demo/entity/Student.java
@@ -2,16 +2,16 @@ package com.example.aiml.entity;
 
 
 public class Student {
-    private int id ;
+    private long id ;
     private String name;
     private String email;
     
     
-    public void setId(int id){
+    public void setId(long id){
         this.id=id;
 
     }
-    public int getId() {
+    public long getId() {
         return id;
         
     }
@@ -28,7 +28,7 @@ public class Student {
         this.email = email;
     }
     
-    public Student(int id, String name, String email) {
+    public Student(long id, String name, String email) {
         this.id = id;
         this.name = name;
         this.email = email;
